refactor: split file-comments tests into a #[path] module
The rule file had grown to 531 lines, over the project's own 500-line
file-length limit, so `cargo run -- lint-extra .` flagged it. Move the
test module to file_comments_tests.rs, matching the five sibling rules
that already split this way, and exclude it from self-linting for the
same `use super::*` reason the others are excluded.
